refactor: share MuPDF disposal helpers
